MinMax Algorithm implemented
